+++ LearningProgress/LearningProgressInterfaces/ViewModels/DisciplineLessonTaskViewModels.cs
@@ -17,7 +17,7 @@ namespace LearningProgressInterfaces.ViewModels
 
         public int Order { get; set; }
 
-        public decimal? MaxBall { get; set; }
+        public double? MaxBall { get; set; }
 
         public string Description { get; set; }
